Make tab bar follow window decoration layout
See https://gitlab.gnome.org/GNOME/libadwaita/-/issues/588

Note that this implementation sort of has a bug where if the layout has
no colon, the window's close button would be on the left, and this puts
close buttons on the right on the tab bar. But this is how it currently
works for all the GNOME apps. Probably you're not supposed to omit the
colon, anyway.